frontend/projects/bookolog-ui-kit/src/lib/ui-progress-bar/components/progress-bar/progress-bar.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { ProgressBarComponent } from './progress-bar.component';

describe('ProgressBarComponent', () => {
  let component: ProgressBarComponent;
  let fixture: ComponentFixture<ProgressBarComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ProgressBarComponent],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(ProgressBarComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  describe('value accessor', () => {
    it('should return value', () => {
      component.value = 80;

      expect(component.value).toEqual(80);
    });

    it('should return 100', () => {
      component.value = 800;

      expect(component.value).toEqual(100);
    });
  });
});
